WIP: refactor hook functions
